(maint) Fix installation failures
When running the initial puppet run to install everything, on a
newly created compiler puppet fails to write the token and bolt-project
files because the /root/.puppetlabs/ dirs don't exist.

This commit adds resources to the init manifest for ensuring the parent
dirs of the token and project files exist before attempting to create
the files.